+++ packages/session/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`session`](https://github.com/remix-run/remix/tree/main/packages/session). It follows [semantic versioning](https://semver.org/).
 
+## Unreleased
+
+- Always delete the original session ID when it is regenerated with the `deleteOldSession` option. Intermediate IDs are never saved to storage, so they can't be deleted.
+
 ## v0.4.0 (2025-11-25)
 
 - Add `Session` class. The `createSession` function now returns an instance of the `Session` class.